Updated vector Q Table to an unordered map
In each pair of the Q_Table, the key is the selected state and the value is a floating point vector of Q-Values

UPDATE: Due to hardware limitations, the Q-Learning algorithm will not be implemented for a full sized 7x6 board. Another more effective AI model will be used instead to tackle this.